refactor(core): Move chat workflow generation and generate titles on a separate WF (no-changelog)
