+++ packages/cli/src/modules/chat-hub/chat-hub-workflow.service.ts
@@ -1,0 +1,487 @@
+import { ChatHubConversationModel, ChatSessionId } from '@n8n/api-types';
+import { Logger } from '@n8n/backend-common';
+import {
+	SharedWorkflow,
+	SharedWorkflowRepository,
+	withTransaction,
+	WorkflowEntity,
+	WorkflowRepository,
+} from '@n8n/db';
+import { Service } from '@n8n/di';
+import { EntityManager } from '@n8n/typeorm';
+import {
+	AGENT_LANGCHAIN_NODE_TYPE,
+	CHAT_TRIGGER_NODE_TYPE,
+	IConnections,
+	IExecuteData,
+	INode,
+	INodeCredentials,
+	IRunExecutionData,
+	IWorkflowBase,
+	MEMORY_BUFFER_WINDOW_NODE_TYPE,
+	MEMORY_MANAGER_NODE_TYPE,
+	NodeConnectionTypes,
+	OperationalError,
+} from 'n8n-workflow';
+import { v4 as uuidv4 } from 'uuid';
+
+import { ChatHubMessage } from './chat-hub-message.entity';
+import {
+	CONVERSATION_TITLE_GENERATION_PROMPT,
+	NODE_NAMES,
+	PROVIDER_NODE_TYPE_MAP,
+} from './chat-hub.constants';
+import { MessageRecord } from './chat-hub.types';
+import { getMaxContextWindowTokens } from './context-limits';
+
+@Service()
+export class ChatHubWorkflowService {
+	constructor(
+		private readonly logger: Logger,
+		private readonly workflowRepository: WorkflowRepository,
+		private readonly sharedWorkflowRepository: SharedWorkflowRepository,
+	) {}
+
+	async createChatWorkflow(
+		userId: string,
+		sessionId: ChatSessionId,
+		projectId: string,
+		history: ChatHubMessage[],
+		humanMessage: string,
+		credentials: INodeCredentials,
+		model: ChatHubConversationModel,
+		systemMessage?: string,
+		trx?: EntityManager,
+	): Promise<{ workflowData: IWorkflowBase; executionData: IRunExecutionData }> {
+		return await withTransaction(this.workflowRepository.manager, trx, async (em) => {
+			this.logger.debug(
+				`Creating chat workflow for user ${userId} and session ${sessionId}, provider ${model.provider}`,
+			);
+
+			const { nodes, connections, executionData } = this.prepareChatWorkflow({
+				userId,
+				sessionId,
+				history,
+				humanMessage,
+				credentials,
+				model,
+				systemMessage,
+			});
+
+			const newWorkflow = new WorkflowEntity();
+			newWorkflow.versionId = uuidv4();
+			newWorkflow.name = `Chat ${sessionId}`;
+			newWorkflow.active = false;
+			newWorkflow.nodes = nodes;
+			newWorkflow.connections = connections;
+
+			const workflow = await em.save<WorkflowEntity>(newWorkflow);
+
+			await em.save<SharedWorkflow>(
+				this.sharedWorkflowRepository.create({
+					role: 'workflow:owner',
+					projectId,
+					workflow,
+				}),
+			);
+
+			return {
+				workflowData: workflow,
+				executionData,
+			};
+		});
+	}
+
+	private prepareChatWorkflow({
+		userId,
+		sessionId,
+		history,
+		humanMessage,
+		credentials,
+		model,
+		systemMessage,
+	}: {
+		userId: string;
+		sessionId: ChatSessionId;
+		history: ChatHubMessage[];
+		humanMessage: string;
+		credentials: INodeCredentials;
+		model: ChatHubConversationModel;
+		systemMessage?: string;
+	}) {
+		const chatTriggerNode = this.createChatTriggerNode();
+		const toolsAgentNode = this.createToolsAgentNode(model, systemMessage);
+		const modelNode = this.createModelNode(credentials, model);
+		const memoryNode = this.createMemoryNode(20);
+		const restoreMemoryNode = this.createMemoryRestoreNode(history);
+		const clearMemoryNode = this.clearMemoryNode();
+
+		const nodes: INode[] = [
+			chatTriggerNode,
+			toolsAgentNode,
+			modelNode,
+			memoryNode,
+			restoreMemoryNode,
+			clearMemoryNode,
+		];
+
+		const connections: IConnections = {
+			[NODE_NAMES.CHAT_TRIGGER]: {
+				main: [
+					[{ node: NODE_NAMES.RESTORE_CHAT_MEMORY, type: NodeConnectionTypes.Main, index: 0 }],
+				],
+			},
+			[NODE_NAMES.RESTORE_CHAT_MEMORY]: {
+				main: [[{ node: NODE_NAMES.REPLY_AGENT, type: NodeConnectionTypes.Main, index: 0 }]],
+			},
+			[NODE_NAMES.CHAT_MODEL]: {
+				// eslint-disable-next-line @typescript-eslint/naming-convention
+				ai_languageModel: [
+					[{ node: NODE_NAMES.REPLY_AGENT, type: NodeConnectionTypes.AiLanguageModel, index: 0 }],
+				],
+			},
+			[NODE_NAMES.MEMORY]: {
+				ai_memory: [
+					[
+						{ node: NODE_NAMES.REPLY_AGENT, type: NodeConnectionTypes.AiMemory, index: 0 },
+						{ node: NODE_NAMES.RESTORE_CHAT_MEMORY, type: NodeConnectionTypes.AiMemory, index: 0 },
+						{ node: NODE_NAMES.CLEAR_CHAT_MEMORY, type: NodeConnectionTypes.AiMemory, index: 0 },
+					],
+				],
+			},
+			[NODE_NAMES.REPLY_AGENT]: {
+				main: [
+					[
+						{
+							node: NODE_NAMES.CLEAR_CHAT_MEMORY,
+							type: NodeConnectionTypes.Main,
+							index: 0,
+						},
+					],
+				],
+			},
+		};
+
+		const nodeExecutionStack: IExecuteData[] = [
+			{
+				node: chatTriggerNode,
+				data: {
+					main: [
+						[
+							{
+								json: {
+									sessionId,
+									action: 'sendMessage',
+									chatInput: humanMessage,
+								},
+							},
+						],
+					],
+				},
+				source: null,
+			},
+		];
+
+		const executionData: IRunExecutionData = {
+			startData: {},
+			resultData: {
+				runData: {},
+			},
+			executionData: {
+				contextData: {},
+				metadata: {},
+				nodeExecutionStack,
+				waitingExecution: {},
+				waitingExecutionSource: {},
+			},
+			manualData: {
+				userId,
+			},
+		};
+
+		return { nodes, connections, executionData };
+	}
+
+	async createTitleGenerationWorkflow(
+		userId: string,
+		sessionId: ChatSessionId,
+		projectId: string,
+		humanMessage: string,
+		credentials: INodeCredentials,
+		model: ChatHubConversationModel,
+		trx?: EntityManager,
+	): Promise<{ workflowData: IWorkflowBase; executionData: IRunExecutionData }> {
+		return await withTransaction(this.workflowRepository.manager, trx, async (em) => {
+			this.logger.debug(
+				`Creating title generation workflow for user ${userId} and session ${sessionId}, provider ${model.provider}`,
+			);
+
+			const { nodes, connections, executionData } = this.prepareTitleGenerationWorkflow(
+				userId,
+				sessionId,
+				credentials,
+				model,
+				humanMessage,
+			);
+
+			const newWorkflow = new WorkflowEntity();
+			newWorkflow.versionId = uuidv4();
+			newWorkflow.name = `Chat ${sessionId} (Title Generation)`;
+			newWorkflow.active = false;
+			newWorkflow.nodes = nodes;
+			newWorkflow.connections = connections;
+
+			const workflow = await em.save<WorkflowEntity>(newWorkflow);
+
+			await em.save<SharedWorkflow>(
+				this.sharedWorkflowRepository.create({
+					role: 'workflow:owner',
+					projectId,
+					workflow,
+				}),
+			);
+
+			return {
+				workflowData: workflow,
+				executionData,
+			};
+		});
+	}
+
+	private prepareTitleGenerationWorkflow(
+		userId: string,
+		sessionId: ChatSessionId,
+		credentials: INodeCredentials,
+		model: ChatHubConversationModel,
+		humanMessage: string,
+	) {
+		const chatTriggerNode = this.createChatTriggerNode();
+		const titleGeneratorAgentNode = this.createTitleGeneratorAgentNode();
+		const modelNode = this.createModelNode(credentials, model);
+
+		const nodes: INode[] = [chatTriggerNode, titleGeneratorAgentNode, modelNode];
+
+		const connections: IConnections = {
+			[NODE_NAMES.CHAT_TRIGGER]: {
+				main: [
+					[{ node: NODE_NAMES.TITLE_GENERATOR_AGENT, type: NodeConnectionTypes.Main, index: 0 }],
+				],
+			},
+			[NODE_NAMES.CHAT_MODEL]: {
+				// eslint-disable-next-line @typescript-eslint/naming-convention
+				ai_languageModel: [
+					[
+						{
+							node: NODE_NAMES.TITLE_GENERATOR_AGENT,
+							type: NodeConnectionTypes.AiLanguageModel,
+							index: 0,
+						},
+					],
+				],
+			},
+		};
+
+		const nodeExecutionStack: IExecuteData[] = [
+			{
+				node: chatTriggerNode,
+				data: {
+					main: [
+						[
+							{
+								json: {
+									sessionId,
+									action: 'sendMessage',
+									chatInput: humanMessage,
+								},
+							},
+						],
+					],
+				},
+				source: null,
+			},
+		];
+
+		const executionData: IRunExecutionData = {
+			startData: {},
+			resultData: {
+				runData: {},
+			},
+			executionData: {
+				contextData: {},
+				metadata: {},
+				nodeExecutionStack,
+				waitingExecution: {},
+				waitingExecutionSource: {},
+			},
+			manualData: {
+				userId,
+			},
+		};
+		return { nodes, connections, executionData };
+	}
+
+	private createChatTriggerNode(): INode {
+		return {
+			parameters: {},
+			type: CHAT_TRIGGER_NODE_TYPE,
+			typeVersion: 1.4,
+			position: [0, 0],
+			id: uuidv4(),
+			name: NODE_NAMES.CHAT_TRIGGER,
+			webhookId: uuidv4(),
+		};
+	}
+
+	private createToolsAgentNode(model: ChatHubConversationModel, systemMessage?: string): INode {
+		return {
+			parameters: {
+				promptType: 'define',
+				text: `={{ $('${NODE_NAMES.CHAT_TRIGGER}').item.json.chatInput }}`,
+				options: {
+					enableStreaming: true,
+					maxTokensFromMemory:
+						model.provider !== 'n8n' && model.provider !== 'custom-agent'
+							? getMaxContextWindowTokens(model.provider, model.model)
+							: undefined,
+					systemMessage,
+				},
+			},
+			type: AGENT_LANGCHAIN_NODE_TYPE,
+			typeVersion: 3,
+			position: [600, 0],
+			id: uuidv4(),
+			name: NODE_NAMES.REPLY_AGENT,
+		};
+	}
+
+	private createModelNode(
+		credentials: INodeCredentials,
+		conversationModel: ChatHubConversationModel,
+	): INode {
+		if (conversationModel.provider === 'n8n' || conversationModel.provider === 'custom-agent') {
+			throw new OperationalError('Custom agent workflows do not require a model node');
+		}
+
+		const { provider, model } = conversationModel;
+		const common = {
+			position: [600, 500] as [number, number],
+			id: uuidv4(),
+			name: NODE_NAMES.CHAT_MODEL,
+			credentials,
+			type: PROVIDER_NODE_TYPE_MAP[provider].name,
+			typeVersion: PROVIDER_NODE_TYPE_MAP[provider].version,
+		};
+
+		switch (provider) {
+			case 'openai':
+				return {
+					...common,
+					parameters: {
+						model: { __rl: true, mode: 'list', value: model },
+						options: {},
+					},
+				};
+			case 'anthropic':
+				return {
+					...common,
+					parameters: {
+						model: {
+							__rl: true,
+							mode: 'list',
+							value: model,
+							cachedResultName: model,
+						},
+						options: {},
+					},
+				};
+			case 'google':
+				return {
+					...common,
+					parameters: {
+						model: { __rl: true, mode: 'list', value: model },
+						options: {},
+					},
+				};
+		}
+	}
+
+	private createMemoryNode(contextWindowLength: number): INode {
+		return {
+			parameters: {
+				sessionIdType: 'customKey',
+				sessionKey: `={{ $('${NODE_NAMES.CHAT_TRIGGER}').item.json.sessionId }}`,
+				contextWindowLength,
+			},
+			type: MEMORY_BUFFER_WINDOW_NODE_TYPE,
+			typeVersion: 1.3,
+			position: [480, 208],
+			id: uuidv4(),
+			name: NODE_NAMES.MEMORY,
+		};
+	}
+
+	private createMemoryRestoreNode(history: ChatHubMessage[]): INode {
+		return {
+			parameters: {
+				mode: 'insert',
+				insertMode: 'override',
+				messages: {
+					messageValues: history
+						// Empty messages can't be restored by the memory manager
+						.filter((message) => message.content.length > 0)
+						.map((message) => {
+							const typeMap: Record<string, MessageRecord['type']> = {
+								human: 'user',
+								ai: 'ai',
+								system: 'system',
+							};
+
+							// TODO: Tool messages etc?
+							return {
+								type: typeMap[message.type] || 'system',
+								message: message.content,
+								hideFromUI: false,
+							};
+						}),
+				},
+			},
+			type: MEMORY_MANAGER_NODE_TYPE,
+			typeVersion: 1.1,
+			position: [224, 0],
+			id: uuidv4(),
+			name: NODE_NAMES.RESTORE_CHAT_MEMORY,
+		};
+	}
+
+	private clearMemoryNode(): INode {
+		return {
+			parameters: {
+				mode: 'delete',
+				deleteMode: 'all',
+			},
+			type: MEMORY_MANAGER_NODE_TYPE,
+			typeVersion: 1.1,
+			position: [976, 0],
+			id: uuidv4(),
+			name: NODE_NAMES.CLEAR_CHAT_MEMORY,
+		};
+	}
+
+	private createTitleGeneratorAgentNode(): INode {
+		return {
+			parameters: {
+				promptType: 'define',
+				text: `={{ $('${NODE_NAMES.CHAT_TRIGGER}').item.json.chatInput }}`,
+				options: {
+					enableStreaming: false,
+					systemMessage: CONVERSATION_TITLE_GENERATION_PROMPT,
+				},
+			},
+			type: AGENT_LANGCHAIN_NODE_TYPE,
+			typeVersion: 3,
+			position: [224, 360],
+			id: uuidv4(),
+			name: NODE_NAMES.TITLE_GENERATOR_AGENT,
+		};
+	}
+}
